+++ test/unit/fuzz-tools.test.js
@@ -106,6 +106,7 @@ import * as cardioV104 from '../../lib/cardio-v104.js';
 import * as vascularV105 from '../../lib/vascular-v105.js';
 import * as vteV106 from '../../lib/vte-v106.js';
 import * as eddecisionV107 from '../../lib/eddecision-v107.js';
+import * as traumaV108 from '../../lib/trauma-v108.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -156,6 +157,7 @@ const MODULES = {
   'vascular-v105.js': vascularV105,
   'vte-v106.js': vteV106,
   'eddecision-v107.js': eddecisionV107,
+  'trauma-v108.js': traumaV108,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];